Move helpers from shared-web to main application
