+++ cosmetics-web/app/helpers/user_management_helper.rb
@@ -1,0 +1,5 @@
+module UserManagementHelper
+  def user_account_url
+    ::KeycloakClient.instance.user_account_url
+  end
+end
