--- shared-web/app/helpers/shared/web/application_helper.rb
@@ -1,6 +1,0 @@
-module Shared
-  module Web
-    module ApplicationHelper
-    end
-  end
-end
