--- shared-web/app/helpers/shared/web/user_management_helper.rb
@@ -1,9 +1,0 @@
-module Shared
-  module Web
-    module UserManagementHelper
-      def user_account_url
-        ::KeycloakClient.instance.user_account_url
-      end
-    end
-  end
-end
